issue-2 / вынос графического генератора в отдельный модуль
